yahoo/requirements.txt
yfinance
pandas
numpy
google-cloud-firestore